Add typing for API organisation queries
