+++ client/src/types/api/common.ts
@@ -1,0 +1,19 @@
+export interface AddressInfo {
+    address1: string | null;
+    address2: string | null;
+    city: string;
+    country: string;
+    postcode: string;
+    state: string;
+}
+
+export interface LinkHref {
+    href: string;
+}
+
+export interface PhotoInfo {
+    full: string;
+    large: string;
+    medium: string;
+    small: string;
+}
